backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
xgboost
joblib
python-dotenv